feat(backend): Enable the seller's register and login component to communicate with the backend
Manage the state of the login and register components through Redux.
Use Redux to detect whether the seller login and registration are successful
Change admin login button css style.
Create a Seller model to store seller information in the Mongo database.
Create a sellerCustomerModel to store the relationship information
between the seller and the customer in the Mongo database.
Once logged in or registered, send the accessToken(jwt) to the client.